move package to root dir
