+++ pyproject.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 description = "Honeycomb OpenTelemetry Distro for Python"
 authors = ["Honeycomb <support@honeycomb.io>"]
 readme = "README.md"
-packages = [{include = "src"}]
+packages = [{include = "honeycomb" }]
 
 [tool.poetry.dependencies]
 python = "^3.10"
@@ -20,9 +20,3 @@ pycodestyle = "^2.10.0"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
-
-[tool.pytest.ini_options]
-pythonpath = "src"
-addopts = [
-    "--import-mode=importlib",
-]
